fix(duplex): use initialization string as key
Technically, the xoodyak crate allows for keyed initialization with an empty key, but the spec indicates Cyclist should ignore other parameters if the key is an empty string. Using the initialization key as a string gets us a valid keyed duplex even if the crate aligns with the spec later on.

BREAKING CHANGE: Invalidates all previous Veil messages and signatures.