feat(pbenc): go back to Balloon Hashing
The more I look at Argon2id, the more I think the use of a duplex as a wide block labeling function is totally fine.

The Balloon Hashing paper's proofs model the hash function as a 2:1 compression function random oracle, but if it's unobjectionable for Argon2 it should be fine here.